[ticket/11768] Renamed interfaces
PHPBB3-11768
